Fix bugs found by the full test suite against a live Neon branch
- SleepLog.hours: route through Decimal(str(x)) before assigning to the
  DecimalField. Assigning a raw float lets Django's DecimalField.to_python()
  convert it via Context.create_decimal_from_float(), which preserves the
  float's exact binary value (6.22 -> Decimal('6.2199999999999997...')) and
  fails the decimal_places=2 validator on almost any non-round duration.
- Milestone ingest upsert: look up an existing row by title alone, not
  (project, title) or 'no project'. Omitting project_code on an update was
  creating a spurious duplicate instead of finding the real row.
- DATABASES OPTIONS: add connect_timeout=30. Neon's serverless computes
  suspend after idle and take ~15-20s to cold-start on the next connection,
  which can stack with Render's own free-tier cold start.
- Test fixes: IST-offset timestamps for sleep day-attribution tests (they
  were UTC, silently shifting past the noon cutoff), Decimal-vs-float
  comparisons for pytest.approx, valid-looking placeholder URLs for
  Django's URLValidator, aware datetimes for forced updated_at values.
- docs/SHORTCUTS.md, docs/INGEST_API.md: correct two examples to match
  actual behavior (hours renders as a bare JSON number; milestone lookup
  key is title alone).

+++ core/tests/test_models.py
@@ -68,7 +68,10 @@ class TestSleepLogDerivation:
         bed = datetime.datetime(2026, 9, 6, 23, 30, tzinfo=datetime.timezone.utc)
         wake = datetime.datetime(2026, 9, 7, 6, 30, tzinfo=datetime.timezone.utc)
         log = SleepLog.objects.create(log_date=DAY0, bed_at=bed, wake_at=wake)
-        assert log.hours == pytest.approx(7.0)
+        # hours is a Decimal — compare as float rather than relying on
+        # pytest.approx's fast-path exact-equality short-circuit, which
+        # only happens to work here because 7.0 has no fractional part.
+        assert float(log.hours) == pytest.approx(7.0)
 
     def test_hours_null_when_one_missing(self):
         bed = datetime.datetime(2026, 9, 6, 23, 30, tzinfo=datetime.timezone.utc)
